Refactor `rustdoc` into multiple sub-modules.
